# openapi.yaml
openapi: 3.1.0
info:
  title: Dejima API
  version: "0.1"
  x-api-version: 1
  description: >
    HTTP/WebSocket control surface for the Dejima daemon (`dejimad`). This is the
    same API the CLI/TUI use and the contract third-party apps build on.

    Status: Dejima is in alpha (0.x). The `v1/` path prefix is intended to be
    stable, but fields may change until `v1.0.0`; an integer `api_version`
    (advertised on `/v1/overview`) tracks the client/daemon contract separately
    from the semver tag. This spec is hand-maintained from the route table in
    `internal/api/server.go` and the request/response shapes in
    `internal/api/types.go` — those are the source of truth and the spec may
    drift slightly between releases.

    WebSocket endpoints (PTY session streams) are listed for discoverability but
    cannot be fully modeled in OpenAPI; see the per-path notes and the
    `SessionEnvelope` schema for the wire protocol.
servers:
  - url: http://{host}
    description: Daemon over the tailnet (DEJIMA_HOST, e.g. 100.84.12.7:7273)
    variables:
      host:
        default: 127.0.0.1:7273
security:
  - bearerToken: []
  - {}   # operator listeners (unix socket / tailnet) carry no per-request token
...
paths:
...
  /v1/islands/{name}/events:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    get:
      tags: [islands]
      summary: The island's recent event log (newest first)
      operationId: islandEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Per-island bearer token (DEJIMA_TOKEN) for the autonomy path.
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema: { type: string }
      description: Island name
...
  responses:
    NotFound:
      description: Island, agent, or resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
...
    Event:
      type: object
      properties:
        type: { type: string, description: "e.g. island.created, agent.task-complete" }
        island: { type: string }
        agent: { type: string }
        timestamp: { type: string, format: date-time }
        payload: { type: object, additionalProperties: true }
...